+++ packages/happn-3/test/integration/data/providers/loki.js
@@ -3,6 +3,7 @@ require('../../../__fixtures/utils/test_helper').describe({ timeout: 20000 }, (t
   const testFileName = test.newTestFile();
   const mockLogger = {
     info: test.sinon.stub(),
+    debug: test.sinon.stub(),
     error: test.sinon.stub(),
     warn: test.sinon.stub(),
     trace: test.sinon.stub(),
@@ -75,7 +76,7 @@ require('../../../__fixtures/utils/test_helper').describe({ timeout: 20000 }, (t
   });
 
   async function testRemove(settings) {
-    const lokiProvider = new LokiDataProvider(mockLogger);
+    const lokiProvider = new LokiDataProvider(null, mockLogger);
     lokiProvider.settings = {
       ...{
         filename: testFileName,
@@ -112,7 +113,7 @@ require('../../../__fixtures/utils/test_helper').describe({ timeout: 20000 }, (t
   }
 
   async function testMerge(settings) {
-    const lokiProvider = new LokiDataProvider(mockLogger);
+    const lokiProvider = new LokiDataProvider(null, mockLogger);
     lokiProvider.settings = {
       ...{
         filename: testFileName,
@@ -130,7 +131,7 @@ require('../../../__fixtures/utils/test_helper').describe({ timeout: 20000 }, (t
   }
 
   async function testIncrement(settings) {
-    const lokiProvider = new LokiDataProvider(mockLogger);
+    const lokiProvider = new LokiDataProvider(null, mockLogger);
     lokiProvider.settings = {
       ...{
         filename: testFileName,
@@ -149,7 +150,7 @@ require('../../../__fixtures/utils/test_helper').describe({ timeout: 20000 }, (t
   }
 
   async function testCount(settings) {
-    const lokiProvider = new LokiDataProvider(mockLogger);
+    const lokiProvider = new LokiDataProvider(null, mockLogger);
     const created = 1634804510343;
     const modified = created;
     lokiProvider.settings = {