+++ packages/happn-3/RELEASES.md
@@ -1074,3 +1074,7 @@
 13.11.2 2023-08-05
 -----------------
 - feat: fixed mesh_gst paths in happner-2
+
+13.12.0 2024-03-27
+-----------------
+- feat: loki provider has new file-size compaction threshold