+++ packages/happn-cluster/RELEASES.md
@@ -368,3 +368,7 @@ non-release 2017-12-14
 12.4.2 2023-08-05
 -----------------
 - feat: fixed mesh_gst paths in happner-2
+
+12.4.3 2024-03-27
+-----------------
+- feat: loki provider has new file-size compaction threshold