+++ packages/happner-2/RELEASES.md
@@ -1064,3 +1064,7 @@ SMC-1645: made error messages more consistent, encapsulating mesh.component.meth
 13.0.3 2023-08-05
 -----------------
   - happner-2: happn-3 auth provider updates
+
+13.1.0 2024-03-27
+-----------------
+- feat: loki provider has new file-size compaction threshold